+++ src/net/command/CommandLoginRealm.java
@@ -15,14 +15,14 @@ public class CommandLoginRealm extends Command {
 		double key = Math.random();
 		int id = this.connection.readInt();
 		if(Server.getRealmList().containsKey(id)) {
-			Server.getRealm(id).getConnectionManager().getConnection().writeByte(PacketID.LOGIN_REALM);
-			Server.getRealm(id).getConnectionManager().getConnection().writeByte(PacketID.LOGIN_NEW_KEY);
+			Server.getRealm(id).getConnectionManager().getConnection().writeShort(PacketID.LOGIN_REALM);
+			Server.getRealm(id).getConnectionManager().getConnection().writeShort(PacketID.LOGIN_NEW_KEY);
 			Server.getRealm(id).getConnectionManager().getConnection().writeDouble(key);
 			Server.getRealm(id).getConnectionManager().getConnection().writeInt(this.player.getAccountId());
 			Server.getRealm(id).getConnectionManager().getConnection().writeInt(this.player.getAccountRank());
 			Server.getRealm(id).getConnectionManager().getConnection().send();
-			this.connection.writeByte(PacketID.LOGIN);
-			this.connection.writeByte(PacketID.LOGIN_REALM_ACCEPTED);
+			this.connection.writeShort(PacketID.LOGIN);
+			this.connection.writeShort(PacketID.LOGIN_REALM_ACCEPTED);
 			this.connection.writeDouble(key);
 			this.connection.writeInt(Server.getRealm(id).getPort());
 			this.connection.send();
