+++ src/net/command/CommandSendRealmList.java
@@ -18,7 +18,7 @@ public class CommandSendRealmList extends Command {
 	}
 
 	public static void sendRealmList(Player player) {
-		player.getConnectionManager().getConnection().writeByte(PacketID.SEND_REALM_LIST);
+		player.getConnectionManager().getConnection().writeShort(PacketID.SEND_REALM_LIST);
 		player.getConnectionManager().getConnection().writeInt(Server.getRealmList().size());
 		for(WorldServer server : Server.getRealmList().values()) {
 			player.getConnectionManager().getConnection().writeInt(server.getRealmID());
